requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
requests>=2.31.0
gspread>=6.1.0
google-auth>=2.29.0
streamlit>=1.38.0
python-dotenv>=1.0.1
pandas>=2.2.0
plotly>=5.22.0
